requirements.txt
alembic==1.5.6
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
BareNecessities==0.2.8
bcrypt==3.2.0
bleach==3.3.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
config==0.5.0.post0
coverage==5.5
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
form==0.0.1
idna==3.1
importlib-metadata==3.7.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
libsass==0.20.1
Mail==2.1.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
module==0.0.4
packaging==20.9
Pillow==8.1.1
pluggy==0.13.1
py==1.10.0
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
scss==0.8.73
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.3
zipp==3.4.1
